Bit of refactoring
